requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.3
fuzzywuzzy==0.17.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
python-Levenshtein==0.12.0
typing-extensions==3.7.2
websockets==6.0
yarl==1.3.0
environs
numexpr
certifi==2018.4.16
requests==2.20.0
urllib3==1.24.2
python-dateutil
cachetools==2.1.0
OSRS-Hiscores